chore: Remove circular dependency in saveAndReturn module
