+++ api.planx.uk/modules/saveAndReturn/controller.ts
@@ -1,73 +1,6 @@
-import { z } from "zod";
-import { ValidatedRequestHandler } from "../../shared/middleware/validate";
 import { resumeApplication } from "./service/resumeApplication";
-import { LowCalSessionData } from "../../types";
 import { findSession, validateSession } from "./service/validateSession";
-import { PaymentRequest } from "@opensystemslab/planx-core/types";
-
-interface ResumeApplicationResponse {
-  message: string;
-  expiryDate?: string | undefined;
-}
-
-export const resumeApplicationSchema = z.object({
-  body: z.object({
-    payload: z.object({
-      teamSlug: z.string(),
-      email: z.string().email(),
-    }),
-  }),
-});
-
-export type ResumeApplication = ValidatedRequestHandler<
-  typeof resumeApplicationSchema,
-  ResumeApplicationResponse
->;
-
-export const resumeApplicationController: ResumeApplication = async (
-  _req,
-  res,
-  next,
-) => {
-  try {
-    const { teamSlug, email } = res.locals.parsedReq.body.payload;
-    const response = await resumeApplication(teamSlug, email);
-    return res.json(response);
-  } catch (error) {
-    return next({
-      error,
-      message: `Failed to send "Resume" email. ${(error as Error).message}`,
-    });
-  }
-};
-
-export interface ValidationResponse {
-  message: string;
-  changesFound: boolean | null;
-  alteredSectionIds?: Array<string>;
-  reconciledSessionData: Omit<LowCalSessionData, "passport">;
-}
-
-interface LockedSessionResponse {
-  message: "Session locked";
-  paymentRequest?: Partial<
-    Pick<PaymentRequest, "id" | "payeeEmail" | "payeeName">
-  >;
-}
-
-export const validateSessionSchema = z.object({
-  body: z.object({
-    payload: z.object({
-      sessionId: z.string(),
-      email: z.string().email(),
-    }),
-  }),
-});
-
-export type ValidateSessionController = ValidatedRequestHandler<
-  typeof validateSessionSchema,
-  ValidationResponse | LockedSessionResponse
->;
+import { ResumeApplication, ValidateSessionController } from "./types";
 
 export const validateSessionController: ValidateSessionController = async (
   _req,
@@ -104,6 +37,23 @@ export const validateSessionController: ValidateSessionController = async (
     return next({
       error,
       message: "Failed to validate session",
+    });
+  }
+};
+
+export const resumeApplicationController: ResumeApplication = async (
+  _req,
+  res,
+  next,
+) => {
+  try {
+    const { teamSlug, email } = res.locals.parsedReq.body.payload;
+    const response = await resumeApplication(teamSlug, email);
+    return res.json(response);
+  } catch (error) {
+    return next({
+      error,
+      message: `Failed to send "Resume" email. ${(error as Error).message}`,
     });
   }
 };
