+++ api.planx.uk/modules/saveAndReturn/routes.ts
@@ -2,12 +2,11 @@ import { Router } from "express";
 
 import {
   resumeApplicationController,
-  resumeApplicationSchema,
   validateSessionController,
-  validateSessionSchema,
 } from "./controller";
 import { sendEmailLimiter } from "../../rateLimit";
 import { validate } from "../../shared/middleware/validate";
+import { resumeApplicationSchema, validateSessionSchema } from "./types";
 
 const router = Router();
